+++ src/static_urls.rs
@@ -60,6 +60,7 @@ pub(crate) const OPERATOR_POLICY_GUIDE_URL: &str =
     "https://rabbitmq.com/docs/parameters#operator-policies";
 pub(crate) const USER_LIMIT_GUIDE_URL: &str = "https://rabbitmq.com/docs/user-limits";
 pub(crate) const PASSWORD_GUIDE_URL: &str = "https://rabbitmq.com/docs/passwords";
+pub(crate) const PLUGIN_GUIDE_URL: &str = "https://rabbitmq.com/docs/plugins";
 pub(crate) const SHOVEL_GUIDE_URL: &str = "https://rabbitmq.com/docs/shovel";
 pub(crate) const FEDERATION_GUIDE_URL: &str = "https://rabbitmq.com/docs/federation";
 pub(crate) const FEDERATED_QUEUES_GUIDE_URL: &str = "https://rabbitmq.com/docs/federated-queues";